[D] Create intermediate `autorisations` in migration
IMPORTANT: This commit needs both the `last_renewal_date` and the
`revocation_date` fields on the `Autorisation` model. It may
therefore only exist between specific commits on the
`refactor-mandats-cherry` branch (whose IDs will vary will rebases...)

To successfully deploy the new version, it is necessary to first deploy
up to this commit, run `python manage.py migrate_mandats`, and only then
deploy the rest of the branch.

I will create a `1.0.0-pre` tag that points here and a `1.0.0` tag that
points to the "end" of the branch, as a reminder of the procedure.